Happiness decays over time
