+++ under_my_scrum_brella/under_my_scrum_brella/settings.py
@@ -52,6 +52,7 @@ MIDDLEWARE = [
     'django.contrib.auth.middleware.AuthenticationMiddleware',
     'django.contrib.messages.middleware.MessageMiddleware',
     'django.middleware.clickjacking.XFrameOptionsMiddleware',
+    'users.middleware.DecayHappinessMiddleware', #Luke - Links custom middleware for happiness
 ]
 
 ROOT_URLCONF = 'under_my_scrum_brella.urls'
