+++ under_my_scrum_brella/under_my_scrum_brella/templates/mypet.html
@@ -128,7 +128,6 @@ window.onload = displayImages;
 
         var happinessBar = document.getElementById("happiness-bar");
         happinessBar.style.width = buddyHappiness + "%";
-
     };
 </script>
 <!--End of Luke code-->
