+++ under_my_scrum_brella/users/models.py
@@ -12,9 +12,12 @@ class UserDetail(models.Model):
     user = models.OneToOneField(User, on_delete=models.CASCADE, primary_key=True)
     buddy_name = models.CharField(max_length=400)
     buddy_type = models.CharField(max_length=400)
-    buddy_happiness = models.FloatField(default=1)
     total_coins = models.IntegerField(default=0)
     total_xp = models.IntegerField(default=0)
+    #Luke - tracks happiness and when it needs to decay
+    buddy_happiness = models.FloatField(default=1)
+    last_happiness_decay_time = models.DateTimeField(auto_now_add=True)
+
 
 class Friend(models.Model):
     user1 = models.ForeignKey(User, on_delete=models.CASCADE, related_name='first_user')
